backend-python/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
anthropic>=0.25.0
openai>=1.30.0
biopython>=1.83
httpx>=0.27.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
pydantic>=2.7.0
